refactor: lens-debate — package.json declares solid-markdown as a runtime dependency for a node-pure type-shape helper
Extracted hasOwnScheme into a new zero-dep @kolu/url-shape leaf package; solid-browser now imports it directly and drops its @kolu/solid-markdown dependency, making node-purity structural instead of subpath-convention-gated.

Agreed by the lowy ⇄ hickey lens debate (finding hickey-5, raised by hickey). Not pushed or merged.
