+++ packages/solid-markdown/package.json
@@ -23,6 +23,7 @@
   },
   "dependencies": {
     "@kolu/html-escape": "workspace:*",
+    "@kolu/url-shape": "workspace:*",
     "dompurify": "^3.2.4",
     "marked": "^18.0.2",
     "marked-alert": "^2.1.2",
